+++ app/Http/Controllers/Api/V1/UserController.php
@@ -9,7 +9,6 @@ use App\Http\Requests\Api\V1\UpdateUserRequest;
 use App\Http\Resources\V1\UserResource;
 use App\Models\User;
 use App\Policies\V1\UserPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class UserController extends ApiController
@@ -29,17 +28,14 @@ class UserController extends ApiController
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-
-            $this->isAble('store', User::class);
+        if ($this->isAble('store', User::class)) {
 
             $model = $request->mappedAttributes();
-
             return new UserResource(User::create($model));
 
-        } catch (AuthorizationException $exception) {
-            return $this->error('You are not authorized to create that resource.', 401);
         }
+
+        return $this->error('You are not authorized to create that resource.', 401);
     }
 
     /**
@@ -65,18 +61,18 @@ class UserController extends ApiController
             $user = User::findOrFail($user_id);
 
             // Policy Defined in AppServiceProvider because of the Folder Structure.
-            $this->isAble('update', $user);
+            if($this->isAble('update', $user)){
 
-            $model = $request->mappedAttributes();
+                $model = $request->mappedAttributes();
+                $user->update($model);
+                return new UserResource($user);
 
-            $user->update($model);
+            }
 
-            return new UserResource($user);
+            return $this->error('You are not authorized to update that resource.', 401);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found', 404);
-        } catch (AuthorizationException $exception) {
-            return $this->error('You are not authorized to update that resource.', 401);
         }
     }
 
@@ -86,13 +82,15 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            $this->isAble('replace', $user);
+            if($this->isAble('update', $user)){
 
-            $model = $request->mappedAttributes();
+                $model = $request->mappedAttributes();
+                $user->update($model);
+                return new UserResource($user);
 
-            $user->update($model);
+            }
 
-            return new UserResource($user);
+            return $this->error('You are not authorized to update that resource.', 401);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found', 404);
@@ -105,13 +103,18 @@ class UserController extends ApiController
     public function destroy($user_id)
     {
         try {
+
             $user = User::findOrFail($user_id);
 
-            $this->isAble('delete', $user);
+            if($this->isAble('delete', $user)){
 
-            $user->delete();
+                $user->delete();
+                return $this->ok('User successfully deleted');
 
-            return $this->ok('User successfully deleted');
+            }
+
+            return $this->error('You are not authorized to delete that resource.', 401);
+
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found', 404);
         }
